refactor: improve file explorer organization
- Split FileExplorer into smaller components (Header, FileList)
- Created shared types and utils
- Improved type safety with better type definitions
- Better organization with proper folder structure
- Added index file for exports
