+++ apps/web/src/components/file-explorer/FileList.tsx
@@ -1,0 +1,16 @@
+import { FileListProps } from './types';
+import { FileTreeNode } from './FileTreeNode';
+
+export function FileList({ files, onContextMenu }: FileListProps) {
+  return (
+    <div className="p-2">
+      {files.map((file, index) => (
+        <FileTreeNode 
+          key={file.name + index} 
+          node={file}
+          onContextMenu={onContextMenu}
+        />
+      ))}
+    </div>
+  );
+}
