+++ apps/web/src/components/file-explorer/index.ts
@@ -1,0 +1,7 @@
+export { FileExplorer } from './FileExplorer';
+export { FileDialog } from './FileDialog';
+export { FileTreeNode } from './FileTreeNode';
+export { Header } from './Header';
+export { FileList } from './FileList';
+export * from './types';
+export * from './utils';
