+++ apps/web/src/components/file-explorer/types.ts
@@ -5,17 +5,21 @@ export interface FileNode {
   expanded?: boolean;
 }
 
+export type ContextMenuType = 'file' | 'folder' | 'background';
+
 export interface ContextMenuState {
   show: boolean;
   x: number;
   y: number;
-  type: 'file' | 'folder' | 'background' | null;
+  type: ContextMenuType | null;
   targetPath?: string;
 }
 
+export type DialogType = 'new-file' | 'new-folder' | 'rename' | null;
+
 export interface DialogState {
   show: boolean;
-  type: 'new-file' | 'new-folder' | 'rename' | null;
+  type: DialogType;
   path?: string;
   defaultValue?: string;
 }
@@ -25,4 +29,14 @@ export interface MenuItem {
   icon: string;
   onClick: () => void;
   divider?: boolean;
+}
+
+export interface FileListProps {
+  files: FileNode[];
+  onContextMenu: (e: React.MouseEvent, type: ContextMenuType, path?: string) => void;
+}
+
+export interface HeaderProps {
+  title: string;
+  subtitle: string;
 }
